Don't summarize large diffs (#8)
New Feature: This pull request adds new methods to the `Commenter` class
that allow for adding and removing a summary from the description field
of a PR. It also modifies the `codeReview` function in `src/review.ts`
to use these new methods. The changes are well-implemented and make the
code more modular.