update job name (#68)
<!-- This is an auto-generated comment: release notes by openai -->
### Summary by OpenAI

**Release Notes**

- **Chore:** Update job name from "chatgpt" to "review" in
`.github/workflows/openai-review.yml` and `README.md`.
- **Documentation:** Remove special thanks section from `README.md`.

This pull request updates the name of a job from "chatgpt" to "review"
in the `.github/workflows/openai-review.yml` file and in two places in
the `README.md` file. Additionally, a section on special thanks has been
removed from the `README.md` file. These changes are purely cosmetic and
do not affect the functionality of the code.
<!-- end of auto-generated comment: release notes by openai -->