Add option bot_icon (#461)
Modify the icon and image dimensions to be customizable, since the icons
displayed in Coderabbit's Slack notifications are excessively large and
clutter the screen.
<!-- This is an auto-generated comment: release notes by OSS CodeRabbit
-->
### Summary by CodeRabbit

- New Feature: Added a new optional input parameter `bot_icon` to the
GitHub Action workflow. This allows users to customize the icon
displayed in Slack notifications sent by the bot. The default value
remains as the pre-defined image tag, ensuring backward compatibility
for existing users who do not specify this new parameter.
<!-- end of auto-generated comment: release notes by OSS CodeRabbit -->